pyproject.toml
[tool.poetry]
name = "make-it-a-quote-qq"
version = "0.3.1"
description = "@makeitaquote twitter机器人的qq版本"
authors = ["Hyiker <hyikerhu0212@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = {extras = ["websockets"], version = "^2.0.0-beta.4"}
nonebot-adapter-cqhttp = "^2.0.0-beta.1"
requests = "^2.28.1"
pillow = "^9.1.1"
pilmoji = "^2.0.1"
cjk-textwrap = "^0.1.0"
pandas = "^1.4.3"
nonebot-plugin-apscheduler = "^0.1.3"
sqlalchemy = "^1.4.39"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
